small fixes

+++ client/src/components/car/CarItemUser.js
@@ -5,6 +5,14 @@ import { deleteRideFromCar } from "../../actions/carActions";
 import isEmpty from "../../validation/is-empty";
 
 export class CarItemUser extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      showUser: true
+    };
+  }
+
   static propTypes = {
     icon: PropTypes.string.isRequired,
     name: PropTypes.string.isRequired,
@@ -15,27 +23,36 @@ export class CarItemUser extends Component {
   };
 
   deleteRider(carid, rideid) {
-    this.props.deleteRideFromCar(carid, rideid);
+    if (
+      window.confirm("Tem a certeza que deseja retirar da lista este piloto?")
+    ) {
+      this.setState({ showUser: false });
+      this.props.deleteRideFromCar(carid, rideid);
+    }
   }
 
   render() {
     const { icon, name, showDelete, carid, rideid } = this.props;
 
     return (
-      <div className="user-block-2-activities">
-        <div className="user-block-2-active">
-          <i className={icon} /> {name}
-          {showDelete &&
-            !isEmpty(rideid) && (
-              <button
-                type="button"
-                onClick={this.deleteRider.bind(this, carid, rideid)}
-                className="btn btn-danger waves-effect waves-light text-uppercase float-right margin-top--9"
-              >
-                Eliminar
-              </button>
-            )}
-        </div>
+      <div>
+        {this.state.showUser && (
+          <div className="user-block-2-activities">
+            <div className="user-block-2-active">
+              <i className={icon} /> {name}
+              {showDelete &&
+                !isEmpty(rideid) && (
+                  <button
+                    type="button"
+                    onClick={this.deleteRider.bind(this, carid, rideid)}
+                    className="btn btn-danger waves-effect waves-light text-uppercase float-right margin-top--9"
+                  >
+                    Eliminar
+                  </button>
+                )}
+            </div>
+          </div>
+        )}
       </div>
     );
   }
